refactor(storage): centralize db errors and null mappers
