+++ internal/storage/users.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"database/sql"
 	"errors"
-	"strings"
 	"time"
 
 	"github.com/kangheeyong/authgate/internal/storage/storeq"
@@ -199,27 +198,6 @@ func (s *Storage) setUserinfo(ctx context.Context, userinfo *oidc.UserInfo, user
 		}
 	}
 	return nil
-}
-
-// isUniqueViolation checks if a PostgreSQL error is a unique constraint violation.
-func isUniqueViolation(err error, constraintName string) bool {
-	msg := err.Error()
-	return strings.Contains(msg, "23505") || strings.Contains(msg, constraintName)
-}
-
-func nullStringToString(v sql.NullString) string {
-	if !v.Valid {
-		return ""
-	}
-	return v.String
-}
-
-func nullStringToPtr(v sql.NullString) *string {
-	if !v.Valid {
-		return nil
-	}
-	s := v.String
-	return &s
 }
 
 // SetUserStatus sets a user's status directly. For testing and admin operations.
